feat: refactor AppProvider to utilize WagmiWrapper for improved structure and add WagmiWrapper component
