+++ packages/web/providers/AppProvider.tsx
@@ -1,57 +1,27 @@
 "use client"
 
 import "@rainbow-me/rainbowkit/styles.css"
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
-import { connectorsForWallets } from "@rainbow-me/rainbowkit"
-import { WagmiProvider, createConfig, http } from "wagmi"
-import { celoAlfajores } from "wagmi/chains"
 import Layout from "../components/Layout"
-import { injectedWallet } from "@rainbow-me/rainbowkit/wallets"
 import { ThemeProvider } from "../components/theme-provider"
 import RainbowKitWrapper from "./RainbowKitWrapper"
 import { AuthProvider } from "./AuthProvider"
-
-const connectors = connectorsForWallets(
-  [
-    {
-      groupName: "Recommended",
-      wallets: [injectedWallet],
-    },
-  ],
-  {
-    appName: "Celo EU",
-    projectId:
-      process.env.REOWN_PROJECT_ID ?? "044601f65212332475a09bc14ceb3c34",
-  }
-)
-
-const config = createConfig({
-  connectors,
-  chains: [celoAlfajores],
-  transports: {
-    [celoAlfajores.id]: http(),
-  },
-})
-
-const queryClient = new QueryClient()
+import WagmiWrapper from "./WagmiWrapper"
 
 export function AppProvider({ children }: { children: React.ReactNode }) {
   return (
-    <WagmiProvider config={config}>
-      <QueryClientProvider client={queryClient}>
-        <AuthProvider>
-          <ThemeProvider
-            attribute="class"
-            defaultTheme="system"
-            enableSystem
-            disableTransitionOnChange
-          >
-            <RainbowKitWrapper>
-              <Layout>{children}</Layout>
-            </RainbowKitWrapper>
-          </ThemeProvider>
-        </AuthProvider>
-      </QueryClientProvider>
-    </WagmiProvider>
+    <WagmiWrapper>
+      <AuthProvider>
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          enableSystem
+          disableTransitionOnChange
+        >
+          <RainbowKitWrapper>
+            <Layout>{children}</Layout>
+          </RainbowKitWrapper>
+        </ThemeProvider>
+      </AuthProvider>
+    </WagmiWrapper>
   )
 }
